feat: should not be able to delete a non-existent beer style

+++ src/modules/beerStyles/useCases/deleteBeerStyle/DeleteBeerStyleUseCase.ts
@@ -1,6 +1,8 @@
 import { IBeerStyleRepository } from "@modules/beerStyles/repositories/IBeerStylesRepository";
 import { inject, injectable } from "tsyringe";
 
+import { AppError } from "@shared/errors/AppError";
+
 @injectable()
 class DeleteBeerStyleUseCase {
   constructor(
@@ -9,6 +11,12 @@ class DeleteBeerStyleUseCase {
   ) {}
 
   async execute(id: string) {
+    const beerStyleExists = await this.beerStylesRepository.findById(id);
+
+    if (!beerStyleExists) {
+      throw new AppError("This beer style doesn't exists!", 404);
+    }
+
     await this.beerStylesRepository.delete(id);
   }
 }
